+++ app/assets/stylesheets/application.scss
@@ -4,6 +4,5 @@
 }
 @import "font-awesome";
 
-@import "application/typography";
 @import "application/pages/*";
 @import "application/components/*";
