--- app/assets/stylesheets/application/_typography.scss
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Dawning+of+a+New+Day|Unna');
