+++ app/assets/stylesheets/application/components/_form.scss
@@ -6,6 +6,9 @@
   position: relative;
   width: 100%;
   display: block;
+  font-family: 'Montserrat', sans-serif;
+  font-size: 1em;
+  color: #8C95AA;
 }
 
 .input-label {
@@ -14,9 +17,10 @@
   left: 0;
   width: 100%;
   height: calc(100% + 4px);
-  margin: 2px;
   text-align: left;
   overflow-x: hidden;
+  font-family: 'Montserrat', sans-serif;
+  font-size: 1em;
 }
 
 .input-label::before, .input-label::after {
@@ -30,11 +34,11 @@
 }
 
 .input-label::before {
-  border-bottom: 2px solid #223843;
+  border-bottom: 1px dotted #9068BE;
 }
 
 .input-label::after {
-  border-bottom: 3px solid #F40076;
+  border-bottom: 3px solid #9068BE;
   transform: translate3d(-100%, 0, 0);
   transition: transform 0.5s;
 }
@@ -44,9 +48,15 @@
   width: 100%;
   border: none;
   background: none;
+  color: #8C95AA;
+  height: 40px;
+  font-family: 'Montserrat', sans-serif;
+  font-size: 16px;
 
   &:focus {
     outline: none;
+    background-color: #4ddce1;
+    color: #fff;
   }
 
   &:invalid {
@@ -66,11 +76,9 @@
 button {
   margin-top: 60px;
   padding: 10px;
-  font-family: 'Unna', serif;
-  font-size: 1.8rem;
+  font-family: 'Montserrat', sans-serif;
+  font-size: 1.6rem;
   color: #9068BE;
-  background-color: transparent;
-  border: 1px solid #1F2B40;
   padding-left: 60px;
   padding-right: 60px;
   display: inline-block;
@@ -78,8 +86,16 @@ button {
   position: relative;
   z-index: 1;
   cursor: pointer;
-  box-shadow: 0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24);
+  background-color: rgba(255, 255, 255, 0.2);
+  box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
+  border: 1px dotted #fff;
   transition: all 0.5s cubic-bezier(.25,.8,.25,1) ease;
+
+  &:hover {
+    color: #fff;
+    text-shadow: 1px 1px 1px #333333;
+    transition: all 0.5s ease-in-out;
+  }
 }
 
 button:after {
@@ -91,11 +107,16 @@ button:after {
   height: 100%;
   top: 0;
   left: 0;
-  background-color: #ffc0cb;
+  background-color: rgba(255, 255, 255, 0.4);
   z-index: -1;
 }
 
 button:hover::after, button:active::after {
   width: 100%;
   box-shadow: 0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);
+  transition: all 0.5s ease-in-out;
+}
+
+textarea.input-field {
+  height: 70px;
 }
