+++ app/assets/stylesheets/application/components/_mfb.scss
@@ -587,7 +587,7 @@
   // background: rgba(255, 255, 255, 0.8);
   padding: 4px 10px;
   border-radius: 3px;
-  color: #fff;
+  color: #8C95AA;
   font-family: 'Montserrat', sans-serif;
   font-size: 10px;
   pointer-events: none;
