+++ app/assets/stylesheets/application/pages/_contact.scss
@@ -1,5 +1,5 @@
 #contact {
-  background-color: #F2C640;
+  background-color: rgba(255, 215, 0, 0.5);
   width: 100vw;
   min-height: 100vh;
   display: flex;
