+++ app/assets/stylesheets/application/pages/_sphere.scss
@@ -16,7 +16,8 @@
   h3 {
     font-family: 'Montserrat', sans-serif;
     font-size: 1.8rem;
-    color: #03243A;
+    color: #fff;
+    text-shadow: 1px 1px 2px #333333;
   }
 }
 
@@ -101,13 +102,13 @@
 
 .lineContainer p {
   font-family: 'Montserrat', sans-serif;
-  font-size: 1.6;
-  color: #03243A;
+  font-size: 0.8rem;
+  color: #183346;
 }
 
 path.line {
   fill: none;
-  stroke: #03243A;
+  stroke: #183346;
   stroke-miterlimit: 10;
   stroke-width: 2px;
 }
@@ -119,7 +120,7 @@ path.line {
   height: 100vh;
   width: 7vh;
   background-color: rgba(255, 255, 255, 0.2);
-  box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
+  // box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -135,7 +136,7 @@ path.line {
     font-family: 'Montserrat', sans-serif;
     font-weight: bolder;
     font-size: 1rem;
-    color:  #791E94;
+    color: #9068BE;
     // text-shadow: 1px 1px 1px #333333;
     z-index: 100;
