+++ app/controllers/portfolio_items_controller.rb
@@ -1,4 +1,7 @@
 class PortfolioItemsController < ApplicationController
+  def index
+  end
+  
   def show
     @portfolio_item = PortfolioItem.find(params[:id])
   end
