+++ app/views/portfolio_items/index.html.haml
@@ -1,0 +1,2 @@
+#carouselBg
+  = react_component("Carousel", props: {}, prerender: false)
