--- client/app/bundles/ContactForm/startup/registration.jsx
@@ -1,8 +1,0 @@
-import ReactOnRails from 'react-on-rails';
-
-import ContactForm from '../components/ContactForm';
-
-// This is how react_on_rails can see the ContactForm in the browser.
-ReactOnRails.register({
-  ContactForm,
-});
