+++ client/package.json
@@ -28,6 +28,7 @@
     "es5-shim": "^4.5.9",
     "expose-loader": "^0.7.1",
     "gsap": "^1.19.1",
+    "hammerjs": "^2.0.8",
     "howler": "^2.0.3",
     "imports-loader": "^0.7.0",
     "react": "^15.4.1",
